Add files via upload
This script scrapes a range of fundamental and technical factors from Yahoo Finance for all stocks listed on the Nasdaq and in the Russell 1000 (although that can be changed by the user). The script also performs a Monte Carlo simulation to predict future return probabilities and formats the data/performs some calculations to make it more conducive to fundamental analysis and machine learning algorithms. 

Please note that the script may take a few hours to complete if it is being run over hundreds/thousands of different stocks. When it is finished running, you will find the output exported as an excel file in your downloads folder.